As reported by Mathieu SCHROETER and following ideas in a patch by Trent Clarke, I fixed the --build command to better select which script should be added to the control file.
Added functions to the architecture class to test for Unix, MS-Windows, Source.
Allow getting the OS, Processor, and Vendor from the architecture.
Added support for detecting "validate" scripts (validate, validate.sh, validate.bat.)
Detect other Unix shells (csh, tcsh, dash, bash, rbash.)
Avoid adding scripts to source packages (At least not in the control.tar.gz file.) I'm not 100% sure that is correct at this point.
Added a test to make sure that the right set of scripts get added to the control.tar.gz file.
Added a quite exhaustive test for the wpkg_architecture implementation.
Fixed bug in architecture compares (<= and >=) which would not work properly when certain parts where equal.
Removed the "x86_64" from the aliases for amd64 because "_" is not at all allowed in a processor name.
Properly verify whether a processor name does not include invalid characters (as per the vendor function) in case extended names are accepted.
Detect the fact that os may be an empty string immediately instead of waiting to test of the os.
Updated the ChangeLog file.
Updated all copyrights to 2014.