Applied the patch by Trent Clarke so all possible installation trees get checked.
Added a function to compare two trees for equivalency (i.e. exact same packages getting installed, verified by name and version.)
Added a tree_generator class to replace the prepare_tree() function that would not go the entire set of possibilities.
Added a function to test whether a package is marked for installation.
Made the necessary changes to make use of the new generator.