Update fixing test of filename to include extension.
See bug report https://sourceforge.net/p/unigw/discussion/647573/thread/f4ff021d/
by Mathiew Shroeter.
Added a test to ensure that filenames having two periods before the extension are now accepted.
Added a test to ensure that a filename ending with a period is forbidden.
Unfortuantely a filename ending with a space is caught earlier so I could not add that test.