Now decorates the debug binaries, and copies them to the right place.
pkg-explorer now links to the debug wpkg binaries correctly, and copies them into the runtime folder. This makes it easier to debug the GUI in devenv (you can just F5 right in the env--the only caveat is that you have to set the working folder to be $(Outdir) instead the default).