Added pkg-installer binary.
This new binary allows you to install one or more packages, like you are
using pkg-explorer, except it does one thing: installs packages. It is
multi-threaded, so it will display the progress via the ProcessDialog
window. It also allows you to add or delete packages from the
ImportDialog, and you can also pass in packages from the command line.