Cleaned up the code for LogOutput, and moved to separate class.
I want to be able to have a log output window pop-up for the
pkg-installer application. To that end, I needed to decouple the class
and utilize Qt4 signals, instead of passing pointers to pure-virtual
classes and having UI elements tightly coupled. This was just bad
design, and you can tell I whipped it up in a small amount of time when
I wrote this stuff originally. :(