Applied another fix by Mathieu Schroeter to fix the --build when used with --install-prefix.
Properly bumped version to 0.9.10 in CMakeLitsts.txt and changelog.
Changed README to use MarkDown syntax.
Changed the control.info file so it makes use of the internal version of wpkg instead of having to update the version in this file each time.
Deprecated the dev/build-debs and dev/build-dist, one should use dev/build.