Now exits with 0 status if no packages are to be installed.
Formerly, when attempting to install a package from sources.list that is
already installed, the wpkg tool would generate a cryptic error
message. Now the tool checks to see if there are 0 packages to install
(because there were no non-installed packages in the list). It generates
a warning, and exits with 0 status, allowing scripting to continue in
this event.

Also, if the number of pending packages is not the same as the number of
packages specified on the command line, a warning is generated, but the
installation proceeds. In both cases, the user is informed about the new
"--force-reinstall" flag.