Tests now run (except complex_tree_in_repository*).
Some of the tests expected the wrong result, like the
depends_distribution_packages* tests. If you try to install package t2,
which relies on t1, but the distribution is wrong, it should fail (but
for some reason, the test was expecting a pass).

The "check_drive_subst" was failing because you don't want to use the
"export" keyword when setting an env var to be passed into the process.

The "complex_tree_in_repository*" tests are still failing, because there
is a bug in the tree trimming code in libdebpackages. A patch will be
forthcoming once I've figured out what the problem actually is.