Fixed upgrade issue.
The problem was that when the wpkg tool command "upgrade" was initiated,
none of the repositories in the core/sources.list file were being added
to the internal repository list. This, of course, would cause errors.
There was no need to do the "temp_d" where we reversed the sense of the
comparison, since with repositories fully loaded, everything seems to
work correctly.