Now adds a proper source as a repository, instead of just the URI.
When I wrote the code that adds repositories, I didn't realize that the
wpkgar_reository::source method "get_uri()" only returned part of the full URL/URI.
So I moved the source class to the wpkgar namespace in wpkgar.h. Now you pass in a
source either as an uri_filename, or as a source, and the source is now properly parsed.
This explains why when using an http source it would fail to load the remote file list and wrongly conclude that the dependencies weren't available.