MH's patch to fix the broken build on Linux.
My bad for not checking the build on Linux after commit, but it was
late. That's my story, and I'm sticking to it. :) Tests under Windows
coming forthwith.

MH's words:
"Commit 371ca43 broke the build on Linux. The POSIX-confirm variant of
putenv(3) takes a char *. The aforementioned commit removed a const_cast
around std::string::c_str's result. This was a good change as it
uncovered another issue: the environment contained pointers to the
stack, which is not going to go well after the corresponding scope is
left (wpkgar_manager::run_one_script).

"While fixing the code I figured that using std::map reduces code
duplication. On Windows _wputenv_s can be used instead of composing the
“NAME=VALUE” argument for _wputenv. The patch is below. It builds on
Debian Jessie (GCC 4.9.2) and Microsoft Windows (Visual C++ 2015).

"Note that there are more occurences of the “putenv(const_cast<char
*="">(…)>” pattern. Those will have to be fixed separately."