Work in progress.
I am in the process of moving all references to the wpkg installer and
manager (and lock) into a new class called "Manager". This should
control the lifetime of the lock in concert with the lifetime of the
wpkgar_manager and wpkgar_installer classes. This is a big change, but
I'm just committing this (and not pushing) as a checkpoint.