Massive overhaul in handling the libdebpackages interface lifetimes.
Now Manager is a singleton, which controls the lifetimes of all of the
libdebpackages objects (including the lock). The threads control the
lifetime of this object. It comes into being when a thread is started,
and is destroyed after the thread is done (by calling
Manager::Release()). This should allow the application to stay running,
but the lock file gone and other processes to alter the database.