Much more stable code. Fixed numerous errors.
The Manager class was formerly popping the desicion dialog if a lock
file was already existing. Took that out, and moved it back into
MainWindow where GUI stuff belongs. This was causing the dialog to be
shown in a background thread, which is a no-no.

In order to process packages, you have to load them first. I now do this
because they used to be loaded on the InitThread and the object
persisted to the other processing threads. But since the manager object
goes away between operations, you can't rely on that anymore.