Now all thread objects hold onto the manager instance, and the static is a weak_ptr.
Each thread object: InitThread, InstallThread, RemoveThread, etc. retains a shared_ptr to
Manager. When the thread object is deleted, that reference is released.

The Manager class no longer holds onto a shared_ptr of itself--instead, it holds a weak_ptr,
which resets to 0 should the caller of "Manager::WeakInstance()" release the instance it
is holding. This means that threads will control the lifetime, and they will be released
when the thread is done and the object is deleted.