Improved installer so that you can specify the version of the package to install.
When just specifying the package name using "wpkgar_install::add_package()", if multiple
of the same packages but with differing version numbers, the validation code would
add each of the packages explicitly, which of course would fail the check for similar files
on the disk. I added the ability to specify the Version field to narrow the selection down to a single package.

From the command line, the user cannot specify the version to install, so the function only selects the greatest version as explicit.