Corrected errors under Linux.
Sadly, the gcc compiler version 4.8.4 cannot figure out the context of
of an auto in a for() loop. For example:

for( auto idx(0); idx < size(); ++idx )

if size() returned unsigned, the compiler thinks that "idx" is signed.
This doesn't happen with MSVC2013. So I don't know what gives here. So I
went ahead and reverted them all back to the concrete types since the
gcc compiler is having a senior moment.

What is nice about the gcc compiler is that with -Wshadow turned on,
I've discovered a couple of foobars and fixed them.