Work in progress!
I'm in the process of refactoring the wpkgar_install object into smaller
classes according to function. The biggest chunk is the new
installer::dependencies class, which encapsulates all of the dependency
calculation and validation. At some point, wpkgar_install will be merely
a façade that encapsulates and hides the objects that do the work.

Then I will set out on the task of unit test building that exercises
each of these objects.