Work in progress.
I moved the progress_scope stuff to its own header/cpp, thus decoupling
it from wpkgar_install. This will allow me to reuse the code that
publishes progress records. By rights, it ought to be a part of
wpkg_output (which itself needs to be refactored), but for now, I'm
keeping it in the "installer" namespace. Maybe this is a candidate for
inclusion into a "details" folder under "libdebpackages"?