Work in progress.
This is a check-point check in, to get my changes into the SCM. I have
reworked the tests so that there is a common class, wpkg_tools, which
any test can use if you need to build and install packages. This will be
necessary for the unittest_installer unit test (to be written).