Added a cmake option that toggles local "catch.h" file use.
If you turn the new option off, then the cmake command "find()" will be
used to locate a copy of "catch.h" on your system.

For now, it is not advised, since the latest versions refactored out the
"Clara" class I am relying on to parse arguments. I think I need to get
libadvgetopt in this project.